32-Bit support (#202)
* remove SQLType::POINTER and changed type of TypeId::POINTER to uintptr_t part 1
* added HASH data type
* Travis 32 Bit Builds & Uploads
* disabling xcode 9.1 (mislabeled)

+++ src/common/vector_operations/set.cpp
@@ -57,7 +57,7 @@ void VectorOperations::Set(Vector &result, Value value) {
 			templated_set_loop<double>(result, value.value_.double_);
 			break;
 		case TypeId::POINTER:
-			templated_set_loop<uint64_t>(result, value.value_.pointer);
+			templated_set_loop<uintptr_t>(result, value.value_.pointer);
 			break;
 		case TypeId::VARCHAR: {
 			auto str = result.string_heap.AddString(value.str_value);
